move fields to separate file
